Error location detection with Moo needed changing
Newer versions of Moo (see below) put the contructor into the
appropriate package, and don't leave it as part of
"Method::Generate::Constructor", so testing for that package no longer
works.

Instead test for package /^\(eval \d+\)$/ and (eval) as the sub, and
if so, $level++.

The change in Moo that broke the previous fix was:

    commit 84b6760de8224ee3a020207e08c22810530e417e
    Author: Graham Knop <haarg@haarg.org>
    Date:   Thu Apr 14 16:50:30 2016 -0400

        fix croak location for inlined BUILDARGS

included in versions 2.002 or newer.