Type::Tiny::Enum - change ::XS encoding to quote strings
If we have something like Enum[qw(a b)] we would communicate with
Type::Tiny::XS via a string like:

   'Enum[a, b]'

We've already been hitting issues with this (see perl rt #129729) on hyphens
but now also hitting it on spaces (see perl rt #133141).  This commit
changes the encoding for enums for Type::Tiny::XS to:

   'Enum["a", "b"]'

Since we use B::perlstring() to encode this, we should also handle any
arbitrary string.  When we're on an old Type::Tiny::XS though, we'll
continue to use the old encoding, we'll just avoid stringification for
even more things.