Fix a precedence bug with modulo
This is what perl thinks of `($to-$from % 2)`:

    $ perl -MO=Deparse,-p -e '($to-$from % 2)'
    ($to - ($from % 2));
    -e syntax OK

When the `slurpy HashRef` starts at an index greater or equal to 2:
* `$from % 2` is either 0 or 1,
* `$to` is greater or equal to 2,

`($to - ($from % 2))` is always greater than 0, and the condition is
always true, even if the number of arguments for the hashref is odd.

If the `slurpy HashRef` starts at index 0, and there's only one
argument, the check fails, but for the wrong reason: `$from % 2` is
equal to 0, as well as `$to`.

If the `slurpy HashRef` starts at index 1, and there's only one argument
for the hashref, the check also fails, again for the wrong reasons:
`$from % 2` is equal to 1, and so is `$to`.