avoid using Data::Dumper for Perl OBJECT refs
Data::Dumper (at least as of 2.192, which is in the Perl git repo)
doesn't support the new SVt_PVOBJ type, and generates a warning
when it encounters it.  It does actually return something, but
the warning can be quite annoying.

This commit special cases reftypes of 'OBJECT', avoiding use of
Data::Dumper altogether.